IPv6: move to utils6 and fields modules some useful functions
